# Avoid frequently invoking native method Class.getCanonicalName